separate into libs and files for better readability
