+++ README.md
@@ -1,7 +1,6 @@
-# [Angular Libraries Support](https://ngx-libs.com) ⛱️   [![Netlify Status](https://api.netlify.com/api/v1/badges/b9665d5e-fb50-4f81-892c-7f47640abac0/deploy-status)](https://app.netlify.com/sites/ngx-libs/deploys)
+# [Angular Libraries Support](https://ngx-libs.com) ⛱️ [![Netlify Status](https://api.netlify.com/api/v1/badges/b9665d5e-fb50-4f81-892c-7f47640abac0/deploy-status)](https://app.netlify.com/sites/ngx-libs/deploys)
 
 ![Angular Libraries Support](https://github.com/eneajaho/ngx-libs/assets/25394362/97bdd713-4058-4034-b257-94a6602dab69)
-
 
 While ng updating Angular projects, it is important to know if a library has added support for the new Angular version. This project is supposed to solve this problem by providing a list of libraries and their support for different Angular versions.
 
@@ -13,7 +12,8 @@ Filter the list by Angular version or library name. Click on the links to get to
 
 If you want to add a library, please create a pull request with the following changes:
 
-- Add the library to the `libs.data.ts` file
+- Create a new file inside `lib-support` library with the name of the library you want to add.
+- Add the library name to the `libs.data.ts` file
 
 ## How to run locally
 
@@ -22,4 +22,5 @@ If you want to add a library, please create a pull request with the following ch
 - Run `npm start`
 
 ## License
+
 MIT
