+++ libs/lib-support/README.md
@@ -1,0 +1,4 @@
+# libs-lib-support
+
+This library was generated with [Nx](https://nx.dev).
+
