+++ libs/lib-support/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "libs/lib-support",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^16.1.0",
+    "@angular/core": "^16.1.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "sideEffects": false
+}
