+++ libs/lib-support/src/index.ts
@@ -1,0 +1,1 @@
+export { LIBRARY_SUPPORT_DATA } from './libs.data';
