+++ libs/lib-support/src/lib/matero-extensions.ts
@@ -1,0 +1,34 @@
+import { LibrarySupport } from '@models';
+
+export const MATERO_EXTENSIONS: LibrarySupport = {
+  name: '@ng-matero/extensions',
+  npmUrl: 'https://www.npmjs.com/package/@ng-matero/extensions',
+  githubUrl: 'https://github.com/ng-matero/extensions',
+  versionSupport: {
+    '16': {
+      libraryVersion: '16.0.0',
+      support: true,
+      link: 'https://github.com/ng-matero/extensions/releases/tag/v16.0.0',
+    },
+    '15': {
+      libraryVersion: '15.0.0',
+      support: true,
+      link: 'https://github.com/ng-matero/extensions/releases/tag/v15.0.0',
+    },
+    '14': {
+      libraryVersion: '14.0.0',
+      support: true,
+      link: 'https://github.com/ng-matero/extensions/releases/tag/v14.0.0',
+    },
+    '13': {
+      libraryVersion: '13.0.0',
+      support: true,
+      link: 'https://github.com/ng-matero/extensions/releases/tag/v13.0.0',
+    },
+    '12': {
+      libraryVersion: '12.0.0',
+      support: true,
+      link: 'https://github.com/ng-matero/extensions/releases/tag/v12.0.0',
+    },
+  },
+};
