+++ libs/lib-support/src/lib/ng-select.ts
@@ -1,0 +1,54 @@
+import { LibrarySupport } from '@models';
+
+export const NG_SELECT: LibrarySupport = {
+  name: '@ng-select/ng-select',
+  npmUrl: 'https://www.npmjs.com/package/@ng-select/ng-select',
+  githubUrl: 'https://github.com/ng-select/ng-select',
+  versionSupport: {
+    '16': {
+      libraryVersion: '11.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v11.0.0',
+    },
+    '15': {
+      libraryVersion: '10.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v10.0.0',
+    },
+    '14': {
+      libraryVersion: '9.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v9.0.0',
+    },
+    '13': {
+      libraryVersion: '8.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v8.0.0',
+    },
+    '12': {
+      libraryVersion: '7.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v7.0.0',
+    },
+    '11': {
+      libraryVersion: '6.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v6.0.0',
+    },
+    '10': {
+      libraryVersion: '5.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v5.0.0',
+    },
+    '9': {
+      libraryVersion: '4.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v4.0.0',
+    },
+    '8': {
+      libraryVersion: '3.0.0',
+      support: true,
+      link: 'https://github.com/ng-select/ng-select/releases/tag/v3.0.0',
+    },
+  },
+};
