+++ libs/lib-support/src/lib/ngx-bootstrap.ts
@@ -1,0 +1,64 @@
+import { LibrarySupport } from '@models';
+
+export const NGX_BOOTSTAP: LibrarySupport = {
+  name: 'ngx-bootstrap',
+  npmUrl: 'https://www.npmjs.com/package/ngx-bootstrap',
+  githubUrl: 'https://github.com/valor-software/ngx-bootstrap',
+  versionSupport: {
+    '16': {
+      libraryVersion: '11.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v11.0.0',
+    },
+    '15': {
+      libraryVersion: '10.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v10.0.0',
+    },
+    '14': {
+      libraryVersion: '9.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v9.0.0',
+    },
+    '13': {
+      libraryVersion: '8.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v8.0.0',
+    },
+    '12': {
+      libraryVersion: '8.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v8.0.0',
+    },
+    '11': {
+      libraryVersion: '7.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v7.0.0',
+    },
+    '10': {
+      libraryVersion: '6.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v6.0.0',
+    },
+    '9': {
+      libraryVersion: '6.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v6.0.0',
+    },
+    '8': {
+      libraryVersion: '5.6.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v5.6.0',
+    },
+    '7': {
+      libraryVersion: '5.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v5.0.0',
+    },
+    '6': {
+      libraryVersion: '3.0.0',
+      support: true,
+      link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v3.0.0',
+    },
+  },
+};
