+++ libs/lib-support/src/lib/ngx-contextmenu.ts
@@ -1,0 +1,34 @@
+import { LibrarySupport } from '@models';
+
+export const NGX_CONTEXT_MENU: LibrarySupport = {
+  name: '@perfectmemory/ngx-contextmenu',
+  npmUrl: 'https://www.npmjs.com/package/@perfectmemory/ngx-contextmenu',
+  githubUrl: 'https://github.com/PerfectMemory/ngx-contextmenu',
+  versionSupport: {
+    '16': {
+      libraryVersion: '16.0.1',
+      support: true,
+      link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v16.0.1',
+    },
+    '15': {
+      libraryVersion: '15.1.1',
+      support: true,
+      link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v15.1.1',
+    },
+    '14': {
+      libraryVersion: '14.1.0',
+      support: true,
+      link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v14.1.0',
+    },
+    '13': {
+      libraryVersion: '8.1.0',
+      support: true,
+      link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v8.1.0',
+    },
+    '12': {
+      libraryVersion: '6.0.0',
+      support: true,
+      link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v6.0.0',
+    },
+  },
+};
