+++ libs/lib-support/src/lib/ngx-emoji-mart.ts
@@ -1,0 +1,34 @@
+import { LibrarySupport } from '@models';
+
+export const NGX_EMOJI_MART: LibrarySupport = {
+  name: '@ctrl/ngx-emoji-mart',
+  npmUrl: 'https://www.npmjs.com/package/@ctrl/ngx-emoji-mart',
+  githubUrl: 'https://github.com/scttcper/ngx-emoji-mart',
+  versionSupport: {
+    '16': {
+      libraryVersion: '9.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v9.0.0',
+    },
+    '15': {
+      libraryVersion: '8.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v8.0.0',
+    },
+    '14': {
+      libraryVersion: '7.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v7.0.0',
+    },
+    '13': {
+      libraryVersion: '6.1.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v6.1.0',
+    },
+    '12': {
+      libraryVersion: '6.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v6.0.0',
+    },
+  },
+};
