+++ libs/lib-support/src/lib/ngx-toastr.ts
@@ -1,0 +1,39 @@
+import { LibrarySupport } from '@models';
+
+export const NGX_TOASTR: LibrarySupport = {
+  name: 'ngx-toastr',
+  npmUrl: 'https://www.npmjs.com/package/ngx-toastr',
+  githubUrl: 'https://github.com/scttcper/ngx-toastr',
+  versionSupport: {
+    '16': {
+      libraryVersion: '17.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v17.0.0',
+    },
+    '15': {
+      libraryVersion: '16.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v16.0.0',
+    },
+    '14': {
+      libraryVersion: '15.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v15.0.0',
+    },
+    '13': {
+      libraryVersion: '14.2.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v14.2.0',
+    },
+    '12': {
+      libraryVersion: '14.0.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v14.0.0',
+    },
+    '11': {
+      libraryVersion: '13.2.0',
+      support: true,
+      link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v13.2.0',
+    },
+  },
+};
