+++ libs/lib-support/src/lib/rx-angular-state.ts
@@ -1,0 +1,34 @@
+import { LibrarySupport } from '@models';
+
+export const RX_ANGULAR_STATE: LibrarySupport = {
+  name: '@rx-angular/state',
+  npmUrl: 'https://www.npmjs.com/package/@rx-angular/state',
+  githubUrl: 'https://github.com/rx-angular/rx-angular',
+  versionSupport: {
+    '16': {
+      libraryVersion: '16.0.0',
+      support: true,
+      link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%4016.0.0',
+    },
+    '15': {
+      libraryVersion: '15.0.0',
+      support: true,
+      link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%4015.0.0',
+    },
+    '14': {
+      libraryVersion: '14.0.0',
+      support: true,
+      link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%4014.0.0',
+    },
+    '13': {
+      libraryVersion: '2.0.0',
+      support: true,
+      link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%402.0.0',
+    },
+    '12': {
+      libraryVersion: '1.0.0',
+      support: true,
+      link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%401.0.0',
+    },
+  },
+};
