+++ libs/lib-support/src/libs.data.ts
@@ -1,0 +1,19 @@
+import { MATERO_EXTENSIONS } from './lib/matero-extensions';
+import { NG_SELECT } from './lib/ng-select';
+import { NGNEAT_CASHEW } from './lib/ngneat-cashew';
+import { NGX_BOOTSTAP } from './lib/ngx-bootstrap';
+import { NGX_CONTEXT_MENU } from './lib/ngx-contextmenu';
+import { NGX_EMOJI_MART } from './lib/ngx-emoji-mart';
+import { NGX_TOASTR } from './lib/ngx-toastr';
+import { RX_ANGULAR_STATE } from './lib/rx-angular-state';
+
+export const LIBRARY_SUPPORT_DATA = [
+  MATERO_EXTENSIONS,
+  NG_SELECT,
+  NGNEAT_CASHEW,
+  NGX_BOOTSTAP,
+  NGX_CONTEXT_MENU,
+  NGX_EMOJI_MART,
+  NGX_TOASTR,
+  RX_ANGULAR_STATE,
+];
