+++ libs/models/README.md
@@ -1,0 +1,4 @@
+# libs-models
+
+This library was generated with [Nx](https://nx.dev).
+
