+++ libs/models/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/libs/models",
+  "lib": {
+    "entryFile": "src/index.ts"
+  }
+}
