+++ libs/models/src/index.ts
@@ -1,0 +1,4 @@
+export {
+  LibraryAngularVersionSupport,
+  LibrarySupport,
+} from './lib/lib-support.interface';
