+++ libs/utils/README.md
@@ -1,0 +1,4 @@
+# libs-utils
+
+This library was generated with [Nx](https://nx.dev).
+
