+++ package.json
@@ -28,19 +28,35 @@
     "@angular-devkit/build-angular": "^16.1.4",
     "@angular-devkit/core": "~16.1.4",
     "@angular-devkit/schematics": "~16.1.4",
+    "@angular-eslint/eslint-plugin": "~16.0.0",
+    "@angular-eslint/eslint-plugin-template": "~16.0.0",
+    "@angular-eslint/template-parser": "~16.0.0",
     "@angular/cli": "~16.1.4",
     "@angular/compiler-cli": "^16.1.5",
+    "@angular/language-service": "^16.1.5",
     "@nx/angular": "16.5.3",
+    "@nx/eslint-plugin": "16.5.3",
+    "@nx/js": "16.5.3",
+    "@nx/linter": "16.5.3",
     "@nx/workspace": "16.5.3",
     "@schematics/angular": "~16.1.4",
     "@types/jasmine": "~4.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.60.1",
+    "@typescript-eslint/parser": "^5.60.1",
+    "eslint": "~8.15.0",
+    "eslint-config-prettier": "8.1.0",
     "jasmine-core": "~4.5.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
+    "ng-packagr": "~16.1.0",
     "nx": "16.5.3",
+    "postcss": "^8.4.5",
+    "postcss-import": "~14.1.0",
+    "postcss-preset-env": "~7.5.0",
+    "postcss-url": "~10.1.3",
     "prettier": "^3.0.0",
     "prettier-plugin-organize-imports": "^3.2.3",
     "typescript": "~4.9.4"
