--- src/app/libs.data.ts
@@ -1,328 +1,0 @@
-import { LibrarySupport } from './lib-support.interface';
-
-export const LIBRARY_SUPPORT_DATA: LibrarySupport[] = [
-  {
-    name: 'ngx-toastr',
-    npmUrl: 'https://www.npmjs.com/package/ngx-toastr',
-    githubUrl: 'https://github.com/scttcper/ngx-toastr',
-    versionSupport: {
-      '16': {
-        libraryVersion: '17.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v17.0.0',
-      },
-      '15': {
-        libraryVersion: '16.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v16.0.0',
-      },
-      '14': {
-        libraryVersion: '15.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v15.0.0',
-      },
-      '13': {
-        libraryVersion: '14.2.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v14.2.0',
-      },
-      '12': {
-        libraryVersion: '14.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v14.0.0',
-      },
-      '11': {
-        libraryVersion: '13.2.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-toastr/releases/tag/v13.2.0',
-      },
-    },
-  },
-  {
-    name: '@rx-angular/state',
-    npmUrl: 'https://www.npmjs.com/package/@rx-angular/state',
-    githubUrl: 'https://github.com/rx-angular/rx-angular',
-    versionSupport: {
-      '16': {
-        libraryVersion: '16.0.0',
-        support: true,
-        link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%4016.0.0',
-      },
-      '15': {
-        libraryVersion: '15.0.0',
-        support: true,
-        link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%4015.0.0',
-      },
-      '14': {
-        libraryVersion: '14.0.0',
-        support: true,
-        link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%4014.0.0',
-      },
-      '13': {
-        libraryVersion: '2.0.0',
-        support: true,
-        link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%402.0.0',
-      },
-      '12': {
-        libraryVersion: '1.0.0',
-        support: true,
-        link: 'https://github.com/rx-angular/rx-angular/releases/tag/state%401.0.0',
-      },
-    },
-  },
-  {
-    name: '@ctrl/ngx-emoji-mart',
-    npmUrl: 'https://www.npmjs.com/package/@ctrl/ngx-emoji-mart',
-    githubUrl: 'https://github.com/scttcper/ngx-emoji-mart',
-    versionSupport: {
-      '16': {
-        libraryVersion: '9.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v9.0.0',
-      },
-      '15': {
-        libraryVersion: '8.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v8.0.0',
-      },
-      '14': {
-        libraryVersion: '7.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v7.0.0',
-      },
-      '13': {
-        libraryVersion: '6.1.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v6.1.0',
-      },
-      '12': {
-        libraryVersion: '6.0.0',
-        support: true,
-        link: 'https://github.com/scttcper/ngx-emoji-mart/releases/tag/v6.0.0',
-      },
-    },
-  },
-  {
-    name: '@ng-matero/extensions',
-    npmUrl: 'https://www.npmjs.com/package/@ng-matero/extensions',
-    githubUrl: 'https://github.com/ng-matero/extensions',
-    versionSupport: {
-      '16': {
-        libraryVersion: '16.0.0',
-        support: true,
-        link: 'https://github.com/ng-matero/extensions/releases/tag/v16.0.0',
-      },
-      '15': {
-        libraryVersion: '15.0.0',
-        support: true,
-        link: 'https://github.com/ng-matero/extensions/releases/tag/v15.0.0',
-      },
-      '14': {
-        libraryVersion: '14.0.0',
-        support: true,
-        link: 'https://github.com/ng-matero/extensions/releases/tag/v14.0.0',
-      },
-      '13': {
-        libraryVersion: '13.0.0',
-        support: true,
-        link: 'https://github.com/ng-matero/extensions/releases/tag/v13.0.0',
-      },
-      '12': {
-        libraryVersion: '12.0.0',
-        support: true,
-        link: 'https://github.com/ng-matero/extensions/releases/tag/v12.0.0',
-      },
-    },
-  },
-  {
-    name: '@ng-select/ng-select',
-    npmUrl: 'https://www.npmjs.com/package/@ng-select/ng-select',
-    githubUrl: 'https://github.com/ng-select/ng-select',
-    versionSupport: {
-      '16': {
-        libraryVersion: '11.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v11.0.0',
-      },
-      '15': {
-        libraryVersion: '10.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v10.0.0',
-      },
-      '14': {
-        libraryVersion: '9.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v9.0.0',
-      },
-      '13': {
-        libraryVersion: '8.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v8.0.0',
-      },
-      '12': {
-        libraryVersion: '7.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v7.0.0',
-      },
-      '11': {
-        libraryVersion: '6.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v6.0.0',
-      },
-      '10': {
-        libraryVersion: '5.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v5.0.0',
-      },
-      '9': {
-        libraryVersion: '4.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v4.0.0',
-      },
-      '8': {
-        libraryVersion: '3.0.0',
-        support: true,
-        link: 'https://github.com/ng-select/ng-select/releases/tag/v3.0.0',
-      },
-    },
-  },
-  {
-    name: '@ngneat/cashew',
-    npmUrl: 'https://www.npmjs.com/package/@ngneat/cashew',
-    githubUrl: 'https://github.com/ngneat/cashew',
-    versionSupport: {
-      '16': {
-        libraryVersion: '3.1.0',
-        support: 'partial',
-        link: 'https://github.com/ngneat/cashew/releases/tag/v3.1.0',
-        note: 'Can be used with Angular 16.0.0',
-      },
-      '15': {
-        libraryVersion: '3.1.0',
-        support: 'partial',
-        link: 'https://github.com/ngneat/cashew/releases/tag/v3.1.0',
-        note: 'Can be used with Angular 15.0.0',
-      },
-      '14': {
-        libraryVersion: '3.1.0',
-        support: true,
-        link: 'https://github.com/ngneat/cashew/releases/tag/v3.1.0',
-      },
-      '13': {
-        libraryVersion: '3.0.0',
-        support: true,
-        link: 'https://github.com/ngneat/cashew/releases/tag/v3.0.0',
-      },
-      '12': {
-        libraryVersion: '2.0.0',
-        support: true,
-        link: 'https://github.com/ngneat/cashew/releases/tag/v2.0.0',
-      },
-      '11': {
-        libraryVersion: '1.3.2',
-        support: 'partial',
-        link: 'https://github.com/ngneat/cashew/releases/tag/v1.3.2',
-        note: 'Can be used with Angular 11.0.0',
-      },
-      '10': {
-        libraryVersion: '1.3.2',
-        support: true,
-        link: 'https://github.com/ngneat/cashew/releases/tag/v1.3.2',
-      },
-    },
-  },
-  {
-    name: 'ngx-bootstrap',
-    npmUrl: 'https://www.npmjs.com/package/ngx-bootstrap',
-    githubUrl: 'https://github.com/valor-software/ngx-bootstrap',
-    versionSupport: {
-      '16': {
-        libraryVersion: '11.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v11.0.0',
-      },
-      '15': {
-        libraryVersion: '10.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v10.0.0',
-      },
-      '14': {
-        libraryVersion: '9.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v9.0.0',
-      },
-      '13': {
-        libraryVersion: '8.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v8.0.0',
-      },
-      '12': {
-        libraryVersion: '8.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v8.0.0',
-      },
-      '11': {
-        libraryVersion: '7.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v7.0.0',
-      },
-      '10': {
-        libraryVersion: '6.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v6.0.0',
-      },
-      '9': {
-        libraryVersion: '6.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v6.0.0',
-      },
-      '8': {
-        libraryVersion: '5.6.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v5.6.0',
-      },
-      '7': {
-        libraryVersion: '5.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v5.0.0',
-      },
-      '6': {
-        libraryVersion: '3.0.0',
-        support: true,
-        link: 'https://github.com/valor-software/ngx-bootstrap/releases/tag/v3.0.0',
-      },
-    },
-  },
-  {
-    name: '@perfectmemory/ngx-contextmenu',
-    npmUrl: 'https://www.npmjs.com/package/@perfectmemory/ngx-contextmenu',
-    githubUrl: 'https://github.com/PerfectMemory/ngx-contextmenu',
-    versionSupport: {
-      '16': {
-        libraryVersion: '16.0.1',
-        support: true,
-        link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v16.0.1',
-      },
-      '15': {
-        libraryVersion: '15.1.1',
-        support: true,
-        link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v15.1.1',
-      },
-      '14': {
-        libraryVersion: '14.1.0',
-        support: true,
-        link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v14.1.0',
-      },
-      '13': {
-        libraryVersion: '8.1.0',
-        support: true,
-        link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v8.1.0',
-      },
-      '12': {
-        libraryVersion: '6.0.0',
-        support: true,
-        link: 'https://github.com/PerfectMemory/ngx-contextmenu/releases/tag/v6.0.0',
-      }
-    },
-  },
-];
