+++ tsconfig.base.json
@@ -1,0 +1,21 @@
+{
+  "compileOnSave": false,
+  "compilerOptions": {
+    "sourceMap": true,
+    "declaration": false,
+    "moduleResolution": "node",
+    "experimentalDecorators": true,
+    "importHelpers": true,
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "target": "ES2022",
+    "module": "ES2022",
+    "lib": ["ES2022", "dom"],
+    "baseUrl": "./",
+    "paths": {
+      "@lib-support": ["libs/lib-support/src/index.ts"],
+      "@models": ["libs/models/src/index.ts"],
+      "@utils": ["libs/utils/src/index.ts"]
+    }
+  }
+}
